ifdef'ing out contents of driver source/header files when not used.
This change wraps driver header/source files in the appropriate
constants to effectively make them empty if that driver is not
used in this build.

The ESP-IDF build system is smart enough to conditionally build
source files depending on project settings (see CMakeLists.txt
in drivers folders). PlatformIO is not, and always builds every
file. This change effectively makes those files no-ops
which serves the same purpose and makes this project able to
compile with PlatformIO (and other build systems).

This addresses, but does not entirely fix, issue #220.